1、FaceNet 特征提取新增人脸对齐 2、人脸检测新5点人脸关键点定位 3、特征提取接口支持多人脸和最佳人脸提取 4、修复人脸框边界精度问题 5、更新 Maven 发布的 groupId

+++ smartjavaai-common/src/main/java/cn/smartjavaai/common/entity/Point.java
@@ -10,27 +10,27 @@ import java.io.Serializable;
  */
 public class Point implements Serializable {
     private static final long serialVersionUID = 1L;
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
